Move `src/sidebar/{util/query-parser => helpers/query-parser}`
The loose distinction between `util/` and `helpers/` is that `util/` contains
more general-purpose code that might be useful outside of the client, whereas
`helpers/` contains more app-specific functions. The query parser is really the
latter as it is closely linked to the annotation filtering logic implemented in
`helpers/filter-annotations.ts`.
