+++ src/sidebar/components/StreamView.tsx
@@ -1,10 +1,10 @@
 import { useCallback, useEffect } from 'preact/hooks';
 
+import { parseHypothesisSearchQuery } from '../helpers/query-parser';
 import { withServices } from '../service-context';
 import type { APIService } from '../services/api';
 import type { ToastMessengerService } from '../services/toast-messenger';
 import { useSidebarStore } from '../store';
-import { parseHypothesisSearchQuery } from '../util/query-parser';
 import ThreadList from './ThreadList';
 import { useRootThread } from './hooks/use-root-thread';
