+++ src/sidebar/helpers/query-parser.ts
