+++ src/sidebar/helpers/thread-annotations.ts
@@ -1,9 +1,9 @@
 import type { Annotation } from '../../types/api';
 import { memoize } from '../util/memoize';
-import { parseFilterQuery } from '../util/query-parser';
 import { buildThread } from './build-thread';
 import type { Thread, BuildThreadOptions } from './build-thread';
 import { filterAnnotations } from './filter-annotations';
+import { parseFilterQuery } from './query-parser';
 import { shouldShowInTab } from './tabs';
 import { sorters } from './thread-sorters';
